Moved comp_tasks.py, cv_tasks.py, move_tasks.py, prequal_tasks.py, test_tasks.py into other_tasks directory within the task_planning directory
