+++ onboard/src/task_planning/task_planning/other_tasks/prequal_tasks.py
@@ -5,7 +5,7 @@ from collections.abc import Coroutine
 
 from rclpy.logging import get_logger
 
-from task_planning import move_tasks
+from task_planning.other_tasks import move_tasks
 from task_planning.interface.cv import CV
 from task_planning.interface.state import State
 from task_planning.task import Task, task
