+++ onboard/src/task_planning/task_planning/other_tasks/test_tasks.py
@@ -1,4 +1,3 @@
-
 from rclpy.logging import get_logger
 from std_msgs.msg import String
